+++ backend/src/main/java/backend/mulkkam/auth/service/KakaoAuthService.java
@@ -4,8 +4,9 @@ import backend.mulkkam.auth.domain.OauthAccount;
 import backend.mulkkam.auth.domain.OauthProvider;
 import backend.mulkkam.auth.dto.KakaoSigninRequest;
 import backend.mulkkam.auth.dto.OauthLoginResponse;
+import backend.mulkkam.auth.infrastructure.OauthJwtTokenHandler;
 import backend.mulkkam.auth.repository.OauthAccountRepository;
-import backend.mulkkam.infrastructure.KakaoRestClient;
+import backend.mulkkam.auth.infrastructure.KakaoRestClient;
 import backend.mulkkam.member.dto.response.KakaoUserInfo;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
